requirements.txt
streamlit>=1.28
langchain==0.1.6
openai>=1.2

#llama-index==0.10.14
#llama-index-core
#llama-index-llms-huggingface
#llama-index-readers-file
#llama-index-llms-ollama
#llama-index-embeddings-huggingface
#llama-index-vector-stores-chroma

#sentence_transformers==2.5.1
google-generativeai==0.4.1
langchain-google-genai==0.0.11
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
langchain-openai==0.0.6
streamlit-feedback
pypdfium2
pypdf
nltk
#pymilvus
pymupdf==1.23.25
faiss-cpu==1.7.4
#chromadb
#aliyun oss
oss2
python-dotenv
#serpapi
google-search-results==2.4.2
numexpr==2.8.6
arxiv==2.1.0
zhipuai==2.0.1.20240423